+++ packages/runtime/src/Suspense.tsx
@@ -126,11 +126,17 @@ interface SuspenseProps {
   [key: string]: any;
 }
 
+function dispatchSuspenseEvent(event: string, id: string) {
+  window.dispatchEvent(new CustomEvent(event, { detail: { id } }));
+}
+const DISPATCH_SUSPENSE_EVENT_STRING = dispatchSuspenseEvent.toString();
+
 export function withSuspense(Component) {
   return (props: SuspenseProps) => {
     const { fallback, id, ...componentProps } = props;
 
-    const [suspenseState, updateSuspenseData] = React.useState({
+
+    const [suspenseState, updateSuspenseData] = React.useState<SuspenseState>({
       id: id,
       data: null,
       done: false,
@@ -151,16 +157,46 @@ export function withSuspense(Component) {
       updateSuspenseData(newState);
     }
 
+
+    // Get SuspenseWrappers from app context
+    const { SuspenseWrappers = [] } = useAppContext();
+
+    // Compose SuspenseWrappers
+    const composeSuspenseWrappers = React.useCallback(
+      (children: React.ReactNode) => {
+        if (!SuspenseWrappers.length) return children;
+
+        return SuspenseWrappers.reduce((WrappedComponent, wrapperConfig) => {
+          const { Wrapper } = wrapperConfig;
+          return <Wrapper id={id}>{WrappedComponent}</Wrapper>;
+        }, children);
+      },
+      [SuspenseWrappers, id],
+    );
+
+    const wrappedComponent = (
+      <>
+        <InlineScript
+          id={`suspense-parse-start-${id}`}
+          script={`(${DISPATCH_SUSPENSE_EVENT_STRING})('ice-suspense-parse-start','${id}');`}
+        />
+        <Component {...componentProps} />
+        <InlineScript
+          id={`suspense-parse-data-${id}`}
+          script={`(${DISPATCH_SUSPENSE_EVENT_STRING})('ice-suspense-parse-data','${id}');`}
+        />
+        <Data id={id} />
+        <InlineScript
+          id={`suspense-parse-end-${id}`}
+          script={`(${DISPATCH_SUSPENSE_EVENT_STRING})('ice-suspense-parse-end','${id}');`}
+        />
+      </>
+    );
+
     return (
       <React.Suspense fallback={fallback || null}>
         <SuspenseContext.Provider value={suspenseState}>
-          <Component {...componentProps} />
-          <Data id={id} />
-          <script
-            dangerouslySetInnerHTML={{
-              __html: `window.dispatchEvent(new CustomEvent('ice-suspense', { detail: { id: ${id ? `'${id}'` : undefined} } }));`,
-            }}
-          />
+          {composeSuspenseWrappers(wrappedComponent)}
         </SuspenseContext.Provider>
       </React.Suspense>
     );
@@ -174,8 +210,25 @@ function Data(props) {
     <script
       id={props.id && `suspense-script-${props.id}`}
       dangerouslySetInnerHTML={{
-        __html: `!function(){window['${LOADER}'] = window['${LOADER}'] || {};window['${LOADER}']['${props.id}'] = ${JSON.stringify(data)}}();`,
+        __html: `!function(){window['${LOADER}'] = window['${LOADER}'] || {};window['${LOADER}']['${props.id}'] = ${JSON.stringify(data)}}();window.dispatchEvent(new CustomEvent('ice-suspense-data', { detail: { id: ${props.id ? `'${props.id}'` : undefined} } }));`,
       }}
+    />
+  );
+}
+
+interface InlineScriptProps {
+  id: string;
+  script: string;
+}
+
+function InlineScript(props: InlineScriptProps) {
+  return (
+    <script
+      id={props.id}
+      dangerouslySetInnerHTML={{
+        __html: props.script,
+      }}
+      suppressHydrationWarning
     />
   );
 }